refactor(procmond): split registration.rs into registration/ submodules
Pure mechanical module extraction with no logic or public-API changes:
- registration/state.rs: RegistrationState enum + Display
- registration/error.rs: RegistrationError + RegistrationResult
- registration/heartbeat.rs: ConnectionStatus, HeartbeatMetrics, HeartbeatMessage
- registration/config.rs: RegistrationConfig + Default
- registration/tests.rs: full unit test module
- registration/mod.rs: RegistrationManager struct + impl, consts, re-exports

All 52 registration unit tests pass; build/clippy/fmt clean.
